+++ myokit/tests/test_simulation_opencl_vs_cvode.py
@@ -14,7 +14,7 @@ import numpy as np
 
 import myokit
 
-from shared import OpenCL_FOUND, OpenCL_DOUBLE_PRECISION, DIR_DATA
+from shared import OpenCL_DOUBLE_PRECISION, DIR_DATA
 from shared import WarningCollector
 
 # Unit testing in Python 2 and 3